requirements.txt
langchain-core==0.3.22
ghapi==1.0.6
streamlit==1.40.2
langchain-google-genai==2.0.6
GitPython==3.1.43